+++ src/ShapeCrawler/ShapeCollection/IMediaShape.cs
@@ -16,7 +16,7 @@ public interface IMediaShape : IShape
     /// <summary>
     ///     Gets MIME type.
     /// </summary>
-    string MIME { get; }
+    string Mime { get; }
     
     /// <summary>
     ///     Gets bytes of video content.
@@ -48,14 +48,14 @@ internal class MediaShape : Shape, IMediaShape
     
     public override bool Removeable => true;
 
-    public string MIME
+    public string Mime
     {
         get
         {
             var p14Media = this.pPicture.NonVisualPictureProperties!.ApplicationNonVisualDrawingProperties!
                 .Descendants<DocumentFormat.OpenXml.Office2010.PowerPoint.Media>().Single();
             var relationship =
-                this.sdkTypedOpenXmlPart.DataPartReferenceRelationships.First(r => r.Id == p14Media.Embed!.Value);
+                this.SdkTypedOpenXmlPart.DataPartReferenceRelationships.First(r => r.Id == p14Media.Embed!.Value);
 
             return relationship.DataPart.ContentType;
         }
@@ -65,7 +65,7 @@ internal class MediaShape : Shape, IMediaShape
     {
         var p14Media = this.pPicture.NonVisualPictureProperties!.ApplicationNonVisualDrawingProperties!
             .Descendants<DocumentFormat.OpenXml.Office2010.PowerPoint.Media>().Single();
-        var relationship = this.sdkTypedOpenXmlPart.DataPartReferenceRelationships.First(r => r.Id == p14Media.Embed!.Value);
+        var relationship = this.SdkTypedOpenXmlPart.DataPartReferenceRelationships.First(r => r.Id == p14Media.Embed!.Value);
         var stream = relationship.DataPart.GetStream();
         var ms = new MemoryStream();
         stream.CopyTo(ms);