+++ src/ShapeCrawler/Texts/TextBox.cs
@@ -6,21 +6,20 @@ using DocumentFormat.OpenXml.Packaging;
 using ShapeCrawler.Positions;
 using ShapeCrawler.ShapeCollection;
 using ShapeCrawler.Shared;
-using ShapeCrawler.Units;
 using SkiaSharp;
 using A = DocumentFormat.OpenXml.Drawing;
 using P = DocumentFormat.OpenXml.Presentation;
 
 namespace ShapeCrawler.Texts;
 
-internal sealed record TextFrame : ITextBox
+internal sealed record TextBox : ITextBox
 {
     private readonly OpenXmlPart sdkTypedOpenXmlPart;
     private readonly OpenXmlElement sdkTextBody;
 
     private TextVerticalAlignment? valignment;
 
-    internal TextFrame(OpenXmlPart sdkTypedOpenXmlPart, OpenXmlElement sdkTextBody)
+    internal TextBox(OpenXmlPart sdkTypedOpenXmlPart, OpenXmlElement sdkTextBody)
     {
         this.sdkTypedOpenXmlPart = sdkTypedOpenXmlPart;
         this.sdkTextBody = sdkTextBody;
@@ -48,35 +47,9 @@ internal sealed record TextFrame : ITextBox
             return sb.ToString();
         }
 
-        set
-        {
-            var paragraphs = this.Paragraphs.ToList();
-            var paragraphWithPortion = paragraphs.FirstOrDefault(p => p.Portions.Any());
-            if (paragraphWithPortion == null)
-            {
-                paragraphWithPortion = paragraphs.First();
-                paragraphWithPortion.Portions.AddText(value);
-            }
-            else
-            {
-                var removingParagraphs = paragraphs.Where(p => p != paragraphWithPortion);
-                foreach (var removingParagraph in removingParagraphs)
-                {
-                    removingParagraph.Remove();
-                }
-
-                paragraphWithPortion.Text = value;
-            }
-
-            if (this.AutofitType == AutofitType.Shrink)
-            {
-                this.ShrinkText(value, paragraphWithPortion);
-            }
-
-            this.ResizeParentShape();
-        }
+        set => this.SetText(value);
     }
-
+    
     public AutofitType AutofitType
     {
         get
@@ -165,7 +138,7 @@ internal sealed record TextFrame : ITextBox
 
     public bool TextWrapped => this.IsTextWrapped();
 
-    public string SDKXPath => new XmlPath(this.sdkTextBody).XPath;
+    public string SdkXPath => new XmlPath(this.sdkTextBody).XPath;
 
     public TextVerticalAlignment VerticalAlignment
     {
@@ -216,7 +189,35 @@ internal sealed record TextFrame : ITextBox
 
         this.valignment = alignmentValue;
     }
+    
+    private void SetText(string value)
+    {
+        var paragraphs = this.Paragraphs.ToList();
+        var portionPara = paragraphs.FirstOrDefault(p => p.Portions.Any());
+        if (portionPara == null)
+        {
+            portionPara = paragraphs.First();
+            portionPara.Portions.AddText(value);
+        }
+        else
+        {
+            var removingParagraphs = paragraphs.Where(p => p != portionPara);
+            foreach (var removingParagraph in removingParagraphs)
+            {
+                removingParagraph.Remove();
+            }
 
+            portionPara.Text = value;
+        }
+
+        if (this.AutofitType == AutofitType.Shrink)
+        {
+            this.ShrinkText(value, portionPara);
+        }
+
+        this.ResizeParentShape();
+    }
+    
     public void ResizeParentShape()
     {
         if (this.AutofitType != AutofitType.Resize)
@@ -228,47 +229,32 @@ internal sealed record TextFrame : ITextBox
         var popularPortion = baseParagraph.Portions.OfType<TextParagraphPortion>().GroupBy(p => p.Font.Size)
             .OrderByDescending(x => x.Count())
             .First().First();
-        var font = popularPortion.Font;
+        var scFont = popularPortion.Font;
 
-        var paint = new SKPaint();
-        paint.TextSize = new Points(font.Size).AsPixels();
-        var fontFamily = font.LatinName == "Calibri Light" ? "Calibri" // for unknown reasons, SkiaSharp uses "Segoe UI" instead of "Calibri Light"
-            : font.LatinName;
-        paint.Typeface = SKTypeface.FromFamilyName(fontFamily);
-        paint.IsAntialias = true;
+        var paint = new SKPaint
+        {
+            IsAntialias = true
+        };
 
         var lMarginPixel = UnitConverter.CentimeterToPixel(this.LeftMargin);
         var rMarginPixel = UnitConverter.CentimeterToPixel(this.RightMargin);
         var tMarginPixel = UnitConverter.CentimeterToPixel(this.TopMargin);
         var bMarginPixel = UnitConverter.CentimeterToPixel(this.BottomMargin);
 
-        var textRect = default(SKRect);
         var text = this.Text.ToUpper();
-        paint.MeasureText(text, ref textRect);
-        var textWidth = (decimal)textRect.Width;
-        var textHeight = (decimal)paint.TextSize;
+        var textWidth = new Text(text, scFont).PxWidth;
+        var textHeight = scFont.Size;
         var shapeSize = new ShapeSize(this.sdkTypedOpenXmlPart, this.sdkTextBody.Ancestors<P.Shape>().First());
         var currentBlockWidth = shapeSize.Width() - lMarginPixel - rMarginPixel;
         var currentBlockHeight = shapeSize.Height() - tMarginPixel - bMarginPixel;
 
         this.UpdateShapeHeight(textWidth, currentBlockWidth, textHeight, tMarginPixel, bMarginPixel, currentBlockHeight, this.sdkTextBody.Parent!);
-        this.UpdateShapeWidthIfNeeded(paint, lMarginPixel, rMarginPixel, this, this.sdkTextBody.Parent!);
+        this.UpdateShapeWidthIfNeeded(lMarginPixel, rMarginPixel, this, this.sdkTextBody.Parent!, scFont);
     }
 
     internal void Draw(SKCanvas slideCanvas, float shapeX, float shapeY)
     {
-        using var paint = new SKPaint();
-        paint.Color = SKColors.Black;
-        var firstPortion = this.Paragraphs.First().Portions.First();
-        paint.TextSize = (float)firstPortion.Font.Size;
-        var typeFace = SKTypeface.FromFamilyName(firstPortion.Font.LatinName);
-        paint.Typeface = typeFace;
-        float leftMarginPx = (float)UnitConverter.CentimeterToPixel(this.LeftMargin);
-        float topMarginPx = (float)UnitConverter.CentimeterToPixel(this.TopMargin);
-        float fontHeightPx = (float)UnitConverter.PointToPixel(16);
-        float x = shapeX + leftMarginPx;
-        float y = shapeY + topMarginPx + fontHeightPx;
-        slideCanvas.DrawText(this.Text, x, y, paint);
+        throw new NotImplementedException();
     }
 
     private decimal GetLeftMargin()
@@ -342,106 +328,39 @@ internal sealed record TextFrame : ITextBox
 
     private void ShrinkText(string newText, IParagraph baseParagraph)
     {
-        var popularPortion = baseParagraph.Portions.GroupBy(p => p.Font!.Size).OrderByDescending(x => x.Count())
-            .First().First();
-        var font = popularPortion.Font;
-
-        var parent = this.sdkTextBody.Parent!;
-        var shapeSize = new ShapeSize(this.sdkTypedOpenXmlPart, parent);
-        var fontSize = GetAdjustedFontSize(newText, font, (int)shapeSize.Width(), (int)shapeSize.Height());
-
-        var paragraphInternal = (Paragraph)baseParagraph;
-        paragraphInternal.SetFontSize(fontSize);
+        var popularFont = baseParagraph.Portions.GroupBy(paraPortion => paraPortion.Font!.Size).OrderByDescending(x => x.Count())
+            .First().First().Font!;
+        var shapeSize = new ShapeSize(this.sdkTypedOpenXmlPart, this.sdkTextBody.Parent!);
+        
+        var text = new Text(newText, popularFont);
+        text.FitInto(shapeSize.Width(), shapeSize.Height());
+        
+        var internalPara = (Paragraph)baseParagraph;
+        internalPara.SetFontSize((int)text.FontSize);
     }
 
     private void UpdateShapeWidthIfNeeded(
-        SKPaint paint,
         decimal lMarginPixel,
         decimal rMarginPixel,
-        TextFrame textFrame,
-        OpenXmlElement parent)
+        TextBox textBox,
+        OpenXmlElement parent,
+        ITextPortionFont font)
     {
-        if (!textFrame.TextWrapped)
+        if (!textBox.TextWrapped)
         {
-            var longerText = textFrame.Paragraphs
+            var longerText = textBox.Paragraphs
                 .Select(x => new { x.Text, x.Text.Length })
                 .OrderByDescending(x => x.Length)
                 .First().Text;
-            var paraTextRect = default(SKRect);
-            var widthInPixels = paint.MeasureText(longerText, ref paraTextRect);
-
-            // SkiaSharp uses 72 Dpi (https://stackoverflow.com/a/69916569/2948684), ShapeCrawler uses 96 Dpi.
-            // 96/72=1.4
-            const double Scale = 1.4;
-            var newWidth = (int)(widthInPixels * Scale) + lMarginPixel + rMarginPixel;
+            
+            var widthInPixels = new Text(longerText, font).PxWidth;
+            
+            var newWidth = (int)(widthInPixels * (decimal)1.4) // SkiaSharp uses 72 Dpi (https://stackoverflow.com/a/69916569/2948684), ShapeCrawler uses 96 Dpi. 96/72 = 1.4 
+                           + lMarginPixel + rMarginPixel;
             new ShapeSize(this.sdkTypedOpenXmlPart, parent).UpdateWidth(newWidth);
         }
     }
-
-    private static int GetAdjustedFontSize(string text, ITextPortionFont font, int shapeWidth, int shapeHeight)
-    {
-        var surface = SKSurface.Create(new SKImageInfo(shapeWidth, shapeHeight));
-        var canvas = surface.Canvas;
-
-        var paint = new SKPaint();
-        var fontSize = font.Size;
-        paint.TextSize = (float)fontSize;
-        paint.Typeface = SKTypeface.FromFamilyName(font.LatinName);
-        paint.IsAntialias = true;
-        const int defaultPaddingSize = 10;
-        const int topBottomPadding = defaultPaddingSize * 2;
-        var wordMaxY = shapeHeight - topBottomPadding;
-
-        var rect = new SKRect(defaultPaddingSize, defaultPaddingSize, shapeWidth - defaultPaddingSize, shapeHeight - defaultPaddingSize);
-
-        var spaceWidth = paint.MeasureText(" ");
-
-        var wordX = rect.Left;
-        var wordY = rect.Top + paint.TextSize;
-
-        var words = text.Split(' ').ToList();
-        for (var i = 0; i < words.Count;)
-        {
-            var word = words[i];
-            var wordWidth = paint.MeasureText(word);
-            if (wordWidth <= rect.Right - wordX)
-            {
-                canvas.DrawText(word, wordX, wordY, paint);
-                wordX += wordWidth + spaceWidth;
-            }
-            else
-            {
-                wordY += paint.FontSpacing;
-                wordX = rect.Left;
-
-                if (wordY > wordMaxY)
-                {
-                    if (paint.TextSize <= 5) // Min reduce font size
-                    {
-                        break;
-                    }
-
-                    paint.TextSize = --paint.TextSize;
-                    wordX = rect.Left;
-                    wordY = rect.Top + paint.TextSize;
-                    i = -1;
-                }
-                else
-                {
-                    wordX += wordWidth + spaceWidth;
-                    canvas.DrawText(word, wordX, wordY, paint);
-                }
-            }
-
-            i++;
-        }
-
-        const int dpi = 96;
-        var points = Math.Round(paint.TextSize * 72 / dpi, 0);
-
-        return (int)points;
-    }
-
+    
     private void UpdateShapeHeight(
         decimal textWidth,
         decimal currentBlockWidth,